DAO e VO ID int64

+++ GerenciadorDeEstoque/DAO/DAO.cs
@@ -51,7 +51,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADU(String email, String nome, String senha, Int32 itemid)
+        public void ADU(String email, String nome, String senha, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -77,7 +77,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDU(String email, String nome, String senha, Int32 itemid)
+        public void RDU(String email, String nome, String senha, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -182,7 +182,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADC(String email, String nome, Int32 itemid)
+        public void ADC(String email, String nome, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -207,7 +207,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDC(String email, String nome, Int32 itemid)
+        public void RDC(String email, String nome, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -236,7 +236,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Venda
 
-        public void IDV(String nome, Int32 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int32 formaPagamento, Int32 CodCliente)
+        public void IDV(String nome, Int64 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int64 formaPagamento, Int64 CodCliente)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -266,7 +266,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADV(String nome, Int32 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int32 formaPagamento, Int32 CodCliente, Int32 itemid)
+        public void ADV(String nome, Int64 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int64 formaPagamento, Int64 CodCliente, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -299,7 +299,7 @@ namespace GerenciadorDeEstoque.DAO
         }
 
 
-        public void RDV(String nome, Int32 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int32 formaPagamento, Int32 CodCliente, Int32 itemid)
+        public void RDV(String nome, Int64 quantidade, String anotacao, Double valorTotal, String nomeCliente, String telefone, Int64 formaPagamento, Int64 CodCliente, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -335,7 +335,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Produto
 
-        public void IDP(Double valor, Int32 quantidade, String nome, String materialUsado, String tipo)
+        public void IDP(Double valor, Int64 quantidade, String nome, String materialUsado, String tipo)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -362,7 +362,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADP(Double valor, Int32 quantidade, String nome, String materialUsado, String tipo, Int32 itemid)
+        public void ADP(Double valor, Int64 quantidade, String nome, String materialUsado, String tipo, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -391,7 +391,7 @@ namespace GerenciadorDeEstoque.DAO
         }
 
 
-        public void RDP(Double valor, Int32 quantidade, String nome, String materialUsado, String tipo, Int32 itemid)
+        public void RDP(Double valor, Int64 quantidade, String nome, String materialUsado, String tipo, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -423,7 +423,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Telefones
 
-        public void IDT(String telefone, Int32 codCliente)
+        public void IDT(String telefone, Int64 codCliente)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -447,7 +447,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADT(String telefone, Int32 codCliente, Int32 itemid)
+        public void ADT(String telefone, Int64 codCliente, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -472,7 +472,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDT(String telefone, Int32 codCliente, Int32 itemid)
+        public void RDT(String telefone, Int64 codCliente, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -501,7 +501,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Vende
 
-        public void IDVENDE(Int32 idvenda, Int32 idproduto)
+        public void IDVENDE(Int32 idvenda, Int64 idproduto)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -525,7 +525,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADVENDE(Int32 idvenda, Int32 idproduto, Int32 itemid)
+        public void ADVENDE(Int64 idvenda, Int64 idproduto, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -550,7 +550,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDVENDE(Int32 idvenda, Int32 idproduto, Int32 itemid)
+        public void RDVENDE(Int64 idvenda, Int64 idproduto, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -602,7 +602,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADTIPO(String nome, Int32 itemid)
+        public void ADTIPO(String nome, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -626,7 +626,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDTIPO(String nome, Int32 itemid)
+        public void RDTIPO(String nome, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -654,7 +654,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Material
 
-        public void IDM(Int32 idmaterial, String nome, Double valor)
+        public void IDM(Int64 idmaterial, String nome, Double valor)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -679,7 +679,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADM(Int32 idmaterial, String nome, Double valor, Int32 itemid)
+        public void ADM(Int64 idmaterial, String nome, Double valor, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -705,7 +705,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDM(Int32 idmaterial, String nome, Double valor, Int32 itemid)
+        public void RDM(Int64 idmaterial, String nome, Double valor, Int64 itemid)
         {
             conexao = new Conexao();
             con = new MySqlConnection();
@@ -734,7 +734,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Material Produto
 
-        public void IDMPRODUTO(Int32 idmaterial, Int32 idproduto)
+        public void IDMPRODUTO(Int64 idmaterial, Int64 idproduto)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -757,7 +757,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADMPRODUTO(Int32 idmaterial, Int32 idproduto, Int32 itemid)
+        public void ADMPRODUTO(Int64 idmaterial, Int64 idproduto, Int64 itemid)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -781,7 +781,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDMPRODUTO(Int32 idmaterial, Int32 idproduto, Int32 itemid)
+        public void RDMPRODUTO(Int64 idmaterial, Int64 idproduto, Int64 itemid)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -832,7 +832,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADR(String tamanho, Double metragem, Int32 itemid)
+        public void ADR(String tamanho, Double metragem, Int64 itemid)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -856,7 +856,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDR(String tamanho, Double metragem, Int32 itemid)
+        public void RDR(String tamanho, Double metragem, Int64 itemid)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -884,7 +884,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Canudo
 
-        public void IDCANUDO(Int32 quantidade, String cor)
+        public void IDCANUDO(Int64 quantidade, String cor)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -907,7 +907,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADCANUDO(Int32 quantidade, String cor, Int32 itemidproduto)
+        public void ADCANUDO(Int64 quantidade, String cor, Int64 itemidproduto)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -931,7 +931,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDCANUDO(Int32 quantidade, String cor, Int32 itemidproduto)
+        public void RDCANUDO(Int64 quantidade, String cor, Int64 itemidproduto)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -959,7 +959,7 @@ namespace GerenciadorDeEstoque.DAO
 
         #region Fita
 
-        public void IDF(Int32 numero, String tipo, String marca, String numeroCor, Double metragem)
+        public void IDF(Int64 numero, String tipo, String marca, String numeroCor, Double metragem)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -985,7 +985,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADF(Int32 numero, String tipo, String marca, String numeroCor, Double metragem, Int32 itemidTipoMaterial)
+        public void ADF(Int64 numero, String tipo, String marca, String numeroCor, Double metragem, Int64 itemidTipoMaterial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1012,7 +1012,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDF(Int32 numero, String tipo, String marca, String numeroCor, Double metragem, Int32 itemidTipoMatreial)
+        public void RDF(Int64 numero, String tipo, String marca, String numeroCor, Double metragem, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1069,7 +1069,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADPAPEL(String tipo, String gramatura, String cor, String tamanho, Int32 itemidTipoMatreial)
+        public void ADPAPEL(String tipo, String gramatura, String cor, String tamanho, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1096,7 +1096,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDPAPEL(String tipo, String gramatura, String cor, String tamanho, Int32 itemidTipoMatreial)
+        public void RDPAPEL(String tipo, String gramatura, String cor, String tamanho, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1151,7 +1151,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADTECIDO(String tipo, String tipoEstampa, Double metragem, Int32 itemidTipoMatreial)
+        public void ADTECIDO(String tipo, String tipoEstampa, Double metragem, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1176,7 +1176,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDTECIDO(String tipo, String tipoEstampa, Double metragem, Int32 itemidTipoMatreial)
+        public void RDTECIDO(String tipo, String tipoEstampa, Double metragem, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1228,7 +1228,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADE(String tamanho, Double espessura, Int32 itemidTipoMatreial)
+        public void ADE(String tamanho, Double espessura, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1252,7 +1252,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDE(String tamanho, Double espessura, Int32 itemidTipoMatreial)
+        public void RDE(String tamanho, Double espessura, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1303,7 +1303,7 @@ namespace GerenciadorDeEstoque.DAO
             }
         }
 
-        public void ADPEROLA(String cor, Double tamanho, Int32 itemidTipoMatreial)
+        public void ADPEROLA(String cor, Double tamanho, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
@@ -1327,7 +1327,7 @@ namespace GerenciadorDeEstoque.DAO
 
         }
 
-        public void RDPEROLA(String cor, Double tamanho, Int32 itemidTipoMatreial)
+        public void RDPEROLA(String cor, Double tamanho, Int64 itemidTipoMatreial)
         {
             con = new MySqlConnection();
             conexao = new Conexao();
